:new: Split utils each into own file
This will help with tests, also help reduce the number of unnecessary utils
