+++ packages/omg/src/utils/error.ts
@@ -1,0 +1,8 @@
+/**
+ * Log a string to stderr.
+ *
+ * @param {String} string The given string to log
+ */
+export default function error(string) {
+  process.stderr.write(`${string}\n`)
+}
