+++ packages/omg/src/utils/index.ts
@@ -1,0 +1,24 @@
+import Docker from 'dockerode-promise'
+
+export { default as appender } from './appender'
+export { default as checkActionInterface } from './checkActionInterface'
+export { default as checkVersion } from './checkVersion'
+export { default as createImageName } from './createImageName'
+export { default as dataTypes } from './dataTypes'
+export { default as doesContainerExist } from './doesContainerExist'
+export { default as error } from './error'
+export { default as exec } from './exec'
+export { default as getForwardPorts } from './getForwardPorts'
+export { default as getHealthPort } from './getHealthPort'
+export { default as getNeededPorts } from './getNeededPorts'
+export { default as getOpenPort } from './getOpenPort'
+export { default as log } from './log'
+export { default as matchEnvironmentCases } from './matchEnvironmentCases'
+export { default as parse } from './parse'
+export { default as readMicroserviceFile } from './readMicroserviceFile'
+export { default as setVal } from './setVal'
+export { default as showVersionCard } from './showVersionCard'
+export { default as sleep } from './sleep'
+export { default as typeCast } from './typeCast'
+
+export const docker = new Docker()
